Update code Update example Remove unneeded anyviews, selfs
