+++ Example/Example/ExampleApp.swift
@@ -1,0 +1,17 @@
+//
+//  ExampleApp.swift
+//  Example
+//
+//  Created by Daniil Manin on 27.09.2021.
+//
+
+import SwiftUI
+
+@main
+struct ExampleApp: App {
+    var body: some Scene {
+        WindowGroup {
+            ContentView()
+        }
+    }
+}
